# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Seu Zuca — Marketplace B2B de Materiais de Construção
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Cadastrar novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Usuário criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login realizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Dados do usuário autenticado
      responses:
        "200":
          description: Usuário autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Não autenticado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Recuperação de senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: E-mail enviado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Redefinir senha
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Senha redefinida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # CATEGORIES
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: Listar categorias
      responses:
        "200":
          description: Lista de categorias
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Criar categoria (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Categoria criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Obter categoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Categoria encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Atualizar categoria (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Categoria atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Excluir categoria (Admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Excluída
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # PRODUCTS
...
  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Adicionar item ao carrinho
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Item adicionado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          description: Quantidade inválida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cart/items/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Atualizar quantidade no carrinho
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Carrinho atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remover item do carrinho
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  # ORDERS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # AUTH SCHEMAS
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        nome:
          type: string
        role:
          type: string
          enum: [buyer, supplier]
        cnpj:
          type: string
        razaoSocial:
          type: string
        nomeFantasia:
          type: string
        telefone:
          type: string
        ramo:
          type: string
        cep:
          type: string
        logradouro:
          type: string
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
        cidade:
          type: string
        estado:
          type: string
      required:
        - email
        - password
        - nome
        - role
        - cnpj
        - razaoSocial
        - telefone
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
      required:
        - token
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        nome:
          type: string
        role:
          type: string
          enum: [admin, buyer, supplier]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        cnpj:
          type: string
        razaoSocial:
          type: string
        nomeFantasia:
          type: string
        telefone:
          type: string
        ramo:
          type: string
        emailVerificado:
          type: boolean
        createdAt:
          type: string
        stripeAccountId:
          type: string
      required:
        - id
        - email
        - nome
        - role
        - status
...
    Category:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        slug:
          type: string
        descricao:
          type: string
        parentId:
          type: integer
        unidadeMedida:
          type: string
        ativo:
          type: boolean
        ordem:
          type: integer
        minimumRule:
          $ref: "#/components/schemas/CategoryMinimumRule"
        children:
          type: array
          items:
            $ref: "#/components/schemas/Category"
      required:
        - id
        - nome
        - slug
        - ativo
    CreateCategoryBody:
      type: object
      properties:
        nome:
          type: string
        slug:
          type: string
        descricao:
          type: string
        parentId:
          type: integer
        unidadeMedida:
          type: string
        ativo:
          type: boolean
        ordem:
          type: integer
      required:
        - nome
        - slug
    CategoryMinimumRule:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        quantidadeMinima:
          type: integer
        multiplo:
          type: integer
        ativo:
          type: boolean
        categoryName:
          type: string
      required:
        - id
        - categoryId
        - quantidadeMinima
        - multiplo
        - ativo
...
    Product:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        slug:
          type: string
        descricao:
          type: string
        sku:
          type: string
        preco:
          type: number
        unidadeMedida:
          type: string
        estoque:
          type: integer
        disponivel:
          type: boolean
        categoryId:
          type: integer
        categoryName:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        imagemPrincipal:
          type: string
        mediaAvaliacao:
          type: number
        totalAvaliacoes:
          type: integer
        prazoFrete:
          type: integer
        quantidadeMinima:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - nome
        - slug
        - disponivel
        - categoryId
        - supplierId
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantidade:
          type: integer
        precoUnitario:
          type: number
        subtotal:
          type: number
        quantidadeMinima:
          type: integer
        valido:
          type: boolean
        mensagemErro:
          type: string
      required:
        - id
        - productId
        - quantidade
        - precoUnitario
        - subtotal
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        valido:
          type: boolean
        erros:
          type: array
          items:
            type: string
      required:
        - items
        - total
        - valido
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantidade:
          type: integer
      required:
        - productId
        - quantidade
    UpdateCartItemBody:
      type: object
      properties:
        quantidade:
          type: integer
      required:
        - quantidade

    # ORDERS
...